Add LivEditMapViewer, a java prog to view where edits are currently been made on a slippy map
It loads the minutely-replicate diffs and the plays back those edits on a slippy map
with about a delay of two minutes. As it has no access to a database, it can
only visualise nodes, as ways and relations don't have the necessary geospacial information
in the diff files.

Caveat: most of this code is the java component jmapviewer which is also checked into the OSM
SVN repository, but this commit doesn't use svn externals to link to it, as it needed some internal changes to it.
This needs cleaning up and consolidating with the jmapviewer component.