Fix to bug 3409 (http://trac.openstreetmap.org/ticket/3409)
I've left in a heap of (disabled) trace code because it's likely someone will want to debug something similar.

In amongst all that there are two lines of code that fix the bug: one that sets selectedIndex to -1, and another that updates the cursor position. Basically with the incorrect index being set, it was partially acting like an autocomplete was being triggered, and then the cursor position was getting set to weird values.

Still some more usability issues remain.