[mod_tile] Allow for per tile style planet-import-complete
If you have several styles served from the same mod_tile, it might be desirable to have
different planet-import-complete dates. This can for example be used if the style sheet changes
and a complete re-render is needed.

It first checks the planet-import-complete file in the tile directory of the style and if the file doesn't
exist checks in the root tile directory.