[osm2pgsql] Replace async I/O in flat-node-cache with posix_fadvise
The idea behind using async I/O was to try and have more than a single I/O operation 
in flight, which can greatly increase performance of raid arrays or SSDs but also has some 
advantages on normal disks. 
 
The glib aio functions however did not work correctly, and were unnecessarily complex. 
 
This patch now replaces all of the async code with a simple call to posix_fadvise(WILLNEED) 
for all of the nodes in a get_nodes_list call. 
 
On my 32Gb SSD, the use of posix_fadvise improves performance of node retrieval from 
3.5k nodes/s to 22k nodes/s under ideal benchmark conditions. In real world usage, 
the speedup is likely to be a lot less, as less I/O can be predicted (most get_node_list calls 
only need to load a single 4k block) and thus less I/O can be parallelised.