[osm2pgsql] Implement a ways_get_list function to retrieve multiple ways at once.
For the postgresql back-end this issues all way requests of a relation 
in a single sql call and then matches the result back to the list of 
ways in osm2pgsql. 
 
This saves round trip time / context switches to postgresql and potentially 
allows postgresql to be more efficient in retrieving the data. 
 
On tests with small extracts, this nearly doubled performance of the 
relation processing stage and showed increases of up to 25% for the complete 
import. On larger imports this speedup is likely smaller, but hopefully 
still noticeable. 
 
This function currently does not yet combine loading of nodes from 
multiple ways in ways_get_list, which might in future provide additional 
speed up.