[mod_tile] Fix previous X-Forwarded-For commit to support chains of X-Forwarded-For
If there are multiple proxies on the path, then X-Forwarded-For is not a single address, but a list
of addresses.

One can now choose whether one wants to use the first entry in the chain, which should be the client address,
or the last entry in the chain.

The first entry of the chain can't necessarily be trusted, as the client can create its own X-Forwarded-For http header
and thus manipulate the IP that the tileserver would use for throttling.

Using the last entry is particularly useful if one has a reverse proxy in front of the tileserver. In that case the
proxy is under ones own control and can be trusted.