use protocol agnostic absolute URLS like //www.openstreetmap.de/karte.html. This fixes problems with HTTPS-Access to the website, where current browsers (I.E. Firefox 27) block unsecure (http) content on secured (https) websites.
Used the protocol on inclusions of external content (there it's this bug fix),
and on links to the website itself (openstreetmap.de), where the protocol currently used should be available in any case.

This commit does NOT do the same for osm.org, wiki.osm.org etc. although it should work there, too.

The map view is not working currently with https due to ilike.openstreetmap.de/ILikeOSM.min.js not being accessible via https. This has to be changed in the configuration (I sent an email to the webmaster@openstreetmap.de address). Nevertheless changed this link, too. This should not change anything for http access and fixes the website code based bug for the https case (except the settings of the subdomains https).